完成剪切器完美设置

+++ gamewindow.h
@@ -54,8 +54,8 @@ private:
     //交付数
 //    int donePieces=0;
     double movingRate=1.0;//传送速度
-    int miningRate=1;//开采速度
-    int cuttingRate=1;//切割速度
+    double miningRate=1;//开采速度
+    double cuttingRate=1;//切割速度
     int flag=0;
     bool isRKeyPressed=false;
     bool isMousePressed = false;
